shared/requirements.txt
redis==5.0.1
aio-pika==9.3.1
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-exporter-prometheus==1.12.0rc1
prometheus-client==0.19.0
pydantic==2.5.1
python-dotenv==1.0.0 